Feat: Serializer, Mixin, Migration files added
- Team 관련 Serializer 추가
- 기존 Team 테이블에서 logo 컬럼 제거
- Team 정보 조회 뷰 추가

+++ teams/models.py
@@ -10,7 +10,7 @@ class Language(models.Model):
         return self.name
     
 class Team(models.Model):
-    id = models.SmallAutoField(primary_key=True)
+    id = models.PositiveBigIntegerField(primary_key=True)
     symbol = models.CharField(max_length=10)
 
     def __str__(self):
